refactor: moved a bunch of logic from App to CameraToolbar, implemented Launch and Relaunch buttons, removed Add Camera button
